refactor(client): globalize design tokens, share auth styles, drop boilerplate
