--- client/src/styles/Landing.module.css
@@ -1,1 +1,0 @@
-/*TODO*/
