--- client/src/styles/NotFound.module.css
@@ -1,1 +1,0 @@
-/*TODO*/
