--- client/src/styles/Register.module.css
@@ -1,276 +1,0 @@
-.page {
-  --bg: #f5f5f4;
-  --bg-dark: #0c0c0e;
-  --card-bg: #ffffff;
-  --card-bg-dark: #141416;
-  --card-border: rgba(0, 0, 0, 0.08);
-  --card-border-dark: rgba(255, 255, 255, 0.07);
-  --text: #111110;
-  --text-dark: #e8e8e6;
-  --text-muted: #78716c;
-  --text-muted-dark: #6b6b68;
-  --input-bg: #f5f5f4;
-  --input-bg-dark: #1c1c1f;
-  --input-border: rgba(0, 0, 0, 0.12);
-  --input-border-dark: rgba(255, 255, 255, 0.08);
-  --input-border-focus: #111110;
-  --input-border-focus-dark: #e8e8e6;
-  --btn-bg: #111110;
-  --btn-bg-dark: #e8e8e6;
-  --btn-text: #ffffff;
-  --btn-text-dark: #111110;
-  --error: #c0392b;
-  --error-bg: #fef2f2;
-  --error-border: rgba(192, 57, 43, 0.2);
-  --mono: ui-monospace, 'JetBrains Mono', Consolas, monospace;
-  --sans: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-
-  position: fixed;
-  inset: 0;
-  overflow-y: auto;
-  background: var(--bg);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 24px;
-}
-
-@media (prefers-color-scheme: dark) {
-  .page {
-    background: var(--bg-dark);
-  }
-}
-
-.card {
-  width: 100%;
-  max-width: 420px;
-  background: var(--card-bg);
-  border: 1px solid var(--card-border);
-  border-radius: 12px;
-  padding: 36px 40px;
-}
-
-@media (prefers-color-scheme: dark) {
-  .card {
-    background: var(--card-bg-dark);
-    border-color: var(--card-border-dark);
-  }
-}
-
-.brand {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  margin-bottom: 28px;
-  font-family: var(--mono);
-  font-size: 14px;
-  font-weight: 500;
-  color: var(--text);
-}
-
-@media (prefers-color-scheme: dark) {
-  .brand {
-    color: var(--text-dark);
-  }
-}
-
-.brandMark {
-  color: var(--text-muted);
-}
-
-@media (prefers-color-scheme: dark) {
-  .brandMark {
-    color: var(--text-muted-dark);
-  }
-}
-
-.brandName {
-  color: var(--text);
-}
-
-@media (prefers-color-scheme: dark) {
-  .brandName {
-    color: var(--text-dark);
-  }
-}
-
-.header {
-  margin-bottom: 28px;
-}
-
-.title {
-  font-family: var(--sans);
-  font-size: 22px;
-  font-weight: 500;
-  color: var(--text);
-  margin: 0 0 4px;
-  letter-spacing: -0.01em;
-}
-
-@media (prefers-color-scheme: dark) {
-  .title {
-    color: var(--text-dark);
-  }
-}
-
-.subtitle {
-  font-family: var(--sans);
-  font-size: 14px;
-  color: var(--text-muted);
-  margin: 0;
-}
-
-@media (prefers-color-scheme: dark) {
-  .subtitle {
-    color: var(--text-muted-dark);
-  }
-}
-
-.errorBanner {
-  background: var(--error-bg);
-  color: var(--error);
-  border: 1px solid var(--error-border);
-  border-radius: 6px;
-  padding: 10px 14px;
-  font-size: 13px;
-  font-family: var(--sans);
-  margin-bottom: 20px;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.field {
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-}
-
-.label {
-  font-family: var(--sans);
-  font-size: 12px;
-  font-weight: 500;
-  color: var(--text-muted);
-  letter-spacing: 0.02em;
-}
-
-@media (prefers-color-scheme: dark) {
-  .label {
-    color: var(--text-muted-dark);
-  }
-}
-
-.input {
-  background: var(--input-bg);
-  border: 1px solid var(--input-border);
-  border-radius: 6px;
-  padding: 10px 12px;
-  font-family: var(--sans);
-  font-size: 14px;
-  color: var(--text);
-  outline: none;
-  transition: border-color 0.15s;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.input::placeholder {
-  color: var(--text-muted);
-  opacity: 0.6;
-}
-
-.input:focus {
-  border-color: var(--input-border-focus);
-}
-
-@media (prefers-color-scheme: dark) {
-  .input {
-    background: var(--input-bg-dark);
-    border-color: var(--input-border-dark);
-    color: var(--text-dark);
-  }
-
-  .input::placeholder {
-    color: var(--text-muted-dark);
-  }
-
-  .input:focus {
-    border-color: var(--input-border-focus-dark);
-  }
-}
-
-.submitBtn {
-  background: var(--btn-bg);
-  color: var(--btn-text);
-  border: none;
-  border-radius: 6px;
-  padding: 11px;
-  font-family: var(--sans);
-  font-size: 14px;
-  font-weight: 500;
-  cursor: pointer;
-  margin-top: 4px;
-  transition: opacity 0.15s;
-  width: 100%;
-}
-
-.submitBtn:hover:not(:disabled) {
-  opacity: 0.85;
-}
-
-.submitBtn:disabled {
-  opacity: 0.45;
-  cursor: not-allowed;
-}
-
-@media (prefers-color-scheme: dark) {
-  .submitBtn {
-    background: var(--btn-bg-dark);
-    color: var(--btn-text-dark);
-  }
-}
-
-.footer {
-  text-align: center;
-  margin: 20px 0 0;
-  color: var(--text-muted);
-  font-family: var(--sans);
-  font-size: 13px;
-}
-
-@media (prefers-color-scheme: dark) {
-  .footer {
-    color: var(--text-muted-dark);
-  }
-}
-
-.link {
-  color: var(--text);
-  text-decoration: underline;
-  text-underline-offset: 2px;
-}
-
-.link:hover {
-  opacity: 0.7;
-}
-
-@media (prefers-color-scheme: dark) {
-  .link {
-    color: var(--text-dark);
-  }
-}
-
-@media (max-width: 480px) {
-  .page {
-    padding: 16px;
-    align-items: flex-start;
-    padding-top: 48px;
-  }
-
-  .card {
-    padding: 28px 24px;
-  }
-}
